+++ packages/react/__tests__/loadable.test.tsx
@@ -1,13 +1,7 @@
 import { render, cleanup, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { afterEach, expect, it, vi } from "vitest";
-import {
-  $computed,
-  createStore,
-  $value,
-  Value,
-  Computed,
-} from "@rippling/core";
+import { $computed, createStore, $value, Value, Computed } from "rippling";
 import {
   StrictMode,
   useEffect,